+++ ui/src/components/card/QasCard.stories.js
@@ -22,28 +22,16 @@ export default {
       description: 'Define a position of the background image.'
     },
 
-    defaultImage: {
-      description: 'Defines a default image when haven\'t a list of images'
-    },
-
-    formMode: {
-      description: 'Changes the component style.'
-    },
-
-    fields: {
-      description: 'Define header value for each `field`.'
+    images: {
+      description: 'List of images.'
     },
 
     gutter: {
       description: 'Gutter space by Quasar convention. `q-col-gutter-{$gutter}`.'
     },
 
-    result: {
-      description: 'Define result value for each `field`.'
-    },
-
-    useActions: {
-      description: 'Shows the actions section.'
+    outlined: {
+      description: 'Changes the component style.'
     },
 
     useHeader: {
@@ -73,7 +61,7 @@ const Template = (args, { argTypes }) => ({
   props: Object.keys(argTypes),
   template:
   `<div class="q-col-gutter-md q-pa-md row">
-    <qas-card useHeader :fields="fields" has-actions :result="result">
+    <qas-card :images="images" use-header>
       <template #default>
         <div>{{ result.name }}</div>
         <div>{{ result.email }}</div>
@@ -84,10 +72,10 @@ const Template = (args, { argTypes }) => ({
         <q-icon color="primary" class="col-grow" name="o_photo_camera" />
       </template>
     </qas-card>
-    <qas-card useHeader :fields="fields" :result="result">
+    <qas-card :images="images" use-header>
       <div>Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.</div>
     </qas-card>
-    <qas-card useHeader :fields="fields" has-actions :result="result">
+    <qas-card :images="images" use-header>
       <template #default>
         <div>{{ result.name }}</div>
         <div>{{ result.email }}</div>
@@ -101,15 +89,12 @@ const Template = (args, { argTypes }) => ({
   </div>`
 })
 
-const fields = {
-  name: { name: 'name', label: 'Full name', type: 'text' },
-  images: [
-    'https://cdn.pixabay.com/photo/2016/10/10/12/54/space-1728314_960_720.jpg',
-    'https://cdn.pixabay.com/photo/2011/12/14/12/21/orion-nebula-11107_960_720.jpg',
-    'https://cdn.pixabay.com/photo/2016/10/10/12/54/space-1728314_960_720.jpg',
-    'https://cdn.pixabay.com/photo/2011/12/14/12/21/orion-nebula-11107_960_720.jpg'
-  ]
-}
+const images = [
+  'https://cdn.pixabay.com/photo/2016/10/10/12/54/space-1728314_960_720.jpg',
+  'https://cdn.pixabay.com/photo/2011/12/14/12/21/orion-nebula-11107_960_720.jpg',
+  'https://cdn.pixabay.com/photo/2016/10/10/12/54/space-1728314_960_720.jpg',
+  'https://cdn.pixabay.com/photo/2011/12/14/12/21/orion-nebula-11107_960_720.jpg'
+]
 
 const result = {
   name: 'John Appleseed',
@@ -122,7 +107,7 @@ const result = {
 
 export const Default = Template.bind({})
 Default.args = {
-  fields,
+  images,
   result
 }
 
@@ -130,7 +115,7 @@ Default.parameters = {
   docs: {
     source: {
       code:
-      '<qas-card :fields="fields" has-actions :result="result" />'
+      '<qas-card :images="images" :result="result" />'
     }
   }
 }